+++ examples/app/routes.js
@@ -1,5 +1,3 @@
-/* istanbul ignore file */
-
 // Sets up the routes.
 module.exports.setup = (app) => {
   /**