+++ examples/app/swaggerDefinition.cjs
@@ -1,5 +1,3 @@
-/* istanbul ignore file */
-
 const host = `http://${process.env.IP}:${process.env.PORT}`;
 
 module.exports = {
@@ -7,8 +5,8 @@ module.exports = {
     // API informations (required)
     title: 'Hello World', // Title (required)
     version: '1.0.0', // Version (required)
-    description: 'A sample API', // Description (optional)
+    description: 'A sample API' // Description (optional)
   },
   host, // Host (optional)
-  basePath: '/', // Base path (optional)
+  basePath: '/' // Base path (optional)
 };