+++ test/files/v2/api_definition.cjs
@@ -1,10 +1,8 @@
-/* istanbul ignore file */
-
 module.exports = {
   info: {
     // API informations (required)
     title: 'Hello World', // Title (required)
     version: '1.0.0', // Version (required)
-    description: 'A sample API', // Description (optional)
-  },
+    description: 'A sample API' // Description (optional)
+  }
 };