+++ test/files/v2/deprecated_routes.js
@@ -1,5 +1,3 @@
-/* istanbul ignore file */
-
 module.exports.setup = function (app) {
   /**
    * @swagger