+++ test/files/v2/wrong-yaml-identation.js
@@ -1,5 +1,3 @@
-/* istanbul ignore file */
-
 module.exports = (app) => {
   /**
    * @swagger