+++ test/files/v2/wrong_definition.cjs
@@ -1,9 +1,7 @@
-/* istanbul ignore file */
-
 const host = `http://${process.env.IP}:${process.env.PORT}`;
 
 module.exports = {
   info: {},
   host, // Host (optional)
-  basePath: '/', // Base path (optional)
+  basePath: '/' // Base path (optional)
 };