+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Changes in the original version can be found from CHANGELOG.old.md file
 - Dropped support for node versions below 20
 - Updated docusaur to latest 2.x version and published docs in github pages
 - Fixed failing unit test and added node versions 20,22,24 to pipeline test matrix
+- Bump eslint from 8 to 9
 - Dependency updates
 
 ## [7.0.0] - version skipped